Refactor files and change output format
